refactor: move configuration para client::configuration
