+++ spec/unit/drasil/client_spec.rb
@@ -24,7 +24,7 @@ RSpec.describe Drasil::Client do
       it "creates a configuration" do
         client = described_class.new(base_url: "https://api.example.com")
 
-        expect(client.config).to be_a(Drasil::Configuration)
+        expect(client.config).to be_a(Drasil::Client::Configuration)
         expect(client.config.base_url).to eq("https://api.example.com")
       end
